+++ src/main/java/dev/ftb/mods/ftbteambases/FTBTeamBases.java
@@ -1,5 +1,6 @@
 package dev.ftb.mods.ftbteambases;
 
+import dev.architectury.platform.Platform;
 import dev.ftb.mods.ftblibrary.config.manager.ConfigManager;
 import dev.ftb.mods.ftbteambases.command.CommandUtils;
 import dev.ftb.mods.ftbteambases.config.ClientConfig;
@@ -9,6 +10,8 @@ import dev.ftb.mods.ftbteambases.data.construction.BaseConstructionManager;
 import dev.ftb.mods.ftbteambases.data.construction.RelocatorTracker;
 import dev.ftb.mods.ftbteambases.data.definition.BaseDefinitionManager;
 import dev.ftb.mods.ftbteambases.data.purging.PurgeManager;
+import dev.ftb.mods.ftbteambases.integration.AutoClaiming;
+import dev.ftb.mods.ftbteambases.integration.FTBChunksIntegration;
 import dev.ftb.mods.ftbteambases.net.SyncBaseTemplatesMessage;
 import dev.ftb.mods.ftbteambases.net.VoidTeamDimensionMessage;
 import dev.ftb.mods.ftbteambases.registry.ModArgumentTypes;
@@ -144,6 +147,7 @@ public class FTBTeamBases {
         if (event.getLevel() instanceof ServerLevel serverLevel) {
             if (serverLevel.dimension() == OVERWORLD) {
                 if (LobbyPregen.maybePregenLobby(serverLevel.getServer())) {
+                    FTBChunksIntegration.maybeAutoClaimLobby(serverLevel);
                     return;
                 }
             }
@@ -151,8 +155,10 @@ public class FTBTeamBases {
             ServerConfig.lobbyDimension().ifPresent(rl -> {
                 if (serverLevel.dimension().equals(rl)) {
                     maybeCreateLobbyFromStructure(serverLevel);
+                    FTBChunksIntegration.maybeAutoClaimLobby(serverLevel);
                 }
             });
+
         }
     }