+++ src/main/java/dev/ftb/mods/ftbteambases/config/ServerConfig.java
@@ -1,6 +1,7 @@
 package dev.ftb.mods.ftbteambases.config;
 
 import dev.ftb.mods.ftblibrary.config.NameMap;
+import dev.ftb.mods.ftblibrary.icon.Color4I;
 import dev.ftb.mods.ftblibrary.snbt.config.*;
 import dev.ftb.mods.ftbteambases.FTBTeamBases;
 import dev.ftb.mods.ftbteambases.worldgen.chunkgen.ChunkGenerators;
@@ -10,10 +11,12 @@ import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.level.ChunkPos;
 import net.minecraft.world.level.GameType;
 import net.minecraft.world.level.Level;
 
 import java.util.Optional;
+import java.util.function.Consumer;
 
 public interface ServerConfig {
     NameMap<GameType> GAME_TYPE_NAME_MAP = NameMap.of(GameType.ADVENTURE, GameType.values()).create();
@@ -80,6 +83,22 @@ public interface ServerConfig {
     IntValue CUSTOM_PORTAL_Y_POS = NETHER.addInt("portal_y_pos", 0)
             .comment("See 'use_custom_portal_y'.");
 
+    SNBTConfig AUTOCLAIMING = CONFIG.addGroup("autoclaiming")
+            .comment("Autoclaim lobby areas (FTB Chunks required)");
+    IntValue LOBBY_RADIUS = AUTOCLAIMING.addInt("lobby_radius", 0, 0, Integer.MAX_VALUE)
+            .comment("Radius in chunks for the lobby area to autoclaim",
+                    "0 = autoclaiming disabled",
+                    "1 = autoclaim just the chunk containing the lobby origin pos",
+                    "2+ = extend the autoclaim distance out by 1 chunk per amount beyond 1");
+    EnumValue<AutoClaimShape> LOBBY_SHAPE = AUTOCLAIMING.addEnum("lobby_shape", AutoClaimShape.NAME_MAP)
+            .comment("Shape to be autoclaimed");
+    StringValue LOBBY_SERVER_TEAM_NAME = AUTOCLAIMING.addString("server_team_name", "Lobby")
+            .comment("The display name for the server team which is used to claim the lobby chunks",
+                    "This name shows up on FTB Chunks mapping");
+    StringValue LOBBY_CLAIM_COLOR = AUTOCLAIMING.addString("lobby_claim_color", "#FF40FF")
+            .comment("The server team color",
+                    "Many color names work, and hex codes in the form '#RRGGBB' are accepted");
+
     static Optional<ResourceLocation> lobbyLocation() {
         try {
             return Optional.of(ResourceLocation.parse(LOBBY_STRUCTURE_LOCATION.get()));
@@ -110,6 +129,49 @@ public interface ServerConfig {
 
     static int getNetherPortalYPos(Player player) {
         return USE_CUSTOM_PORTAL_Y_POS.get() ? CUSTOM_PORTAL_Y_POS.get() : player.blockPosition().getY();
+    }
+
+    static Color4I getLobbyTeamColor() {
+        Color4I teamColor = Color4I.fromString(ServerConfig.LOBBY_CLAIM_COLOR.get());
+        return teamColor.isEmpty() ? Color4I.rgb(0xFF40FF) : teamColor;
+    }
+
+    enum AutoClaimShape {
+        SQUARE("square"),
+        CIRCLE("circle");
+
+        public static final NameMap<AutoClaimShape> NAME_MAP = NameMap.of(SQUARE, values()).id(AutoClaimShape::getId).create();
+
+        private final String shape;
+
+        AutoClaimShape(String id) {
+            this.shape = id;
+        }
+
+        public String getId() {
+            return shape;
+        }
+
+        public void forEachChunk(BlockPos origin, int radius, Consumer<ChunkPos> consumer) {
+            ChunkPos chunk0 = new ChunkPos(origin);
+            BlockPos pos0 = chunk0.getMiddleBlockPosition(0);
+            int blockRadiusSq = (radius << 4) * (radius << 4);
+            switch (radius) {
+                case 0 -> { }
+                case 1 -> consumer.accept(chunk0);
+                default -> {
+                    int r = radius - 1;
+                    for (int cx = chunk0.x - r; cx <= chunk0.x + r; cx++) {
+                        for (int cz = chunk0.z - r; cz <= chunk0.z + r; cz++) {
+                            ChunkPos cp = new ChunkPos(cx, cz);
+                            if (this == SQUARE || cp.getMiddleBlockPosition(0).distSqr(pos0) < blockRadiusSq) {
+                                consumer.accept(cp);
+                            }
+                        }
+                    }
+                }
+            };
+        }
     }
 
     enum FeatureGeneration {