+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [21.1.10]
+
+### Added
+* Added automatic chunk claiming for the lobby area, using FTB Chunks
+  * See the new "autoclaiming" section in `ftbteambases-server.snbt` config file for control of this feature
+  * FTB Teams 2101.1.7 or later required
+
 ## [21.1.9]
 
 ### Fixed